Factor out base class
